backend/requirements.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fuzzywuzzy==0.18.0
joblib==1.5.1
Levenshtein==0.27.1
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytz==2025.2
RapidFuzz==3.13.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
xgboost==3.0.2
psycopg2-binary==2.9.10
imbalanced-learn==0.13.0